refactor: move plugin npm dependencies to a central place
