+++ src/plugins/petpet.ts
@@ -17,8 +17,9 @@
 */
 
 import { ApplicationCommandInputType, ApplicationCommandOptionType, Argument, CommandContext, findOption } from "../api/Commands";
+import { lazyWebpack, makeLazy } from "../utils";
 import { Devs } from "../utils/constants";
-import { lazyWebpack, makeLazy } from "../utils/misc";
+import { getGifEncoder } from "../utils/dependencies";
 import definePlugin from "../utils/types";
 import { filters } from "../webpack";
 
@@ -26,11 +27,6 @@ const DRAFT_TYPE = 0;
 const DEFAULT_DELAY = 20;
 const DEFAULT_RESOLUTION = 128;
 const FRAMES = 10;
-
-// https://github.com/mattdesl/gifenc
-// this lib is way better than gif.js and all other libs, they're all so terrible but this one is nice
-// @ts-ignore ts mad
-const getGifEncoder = makeLazy(() => import("https://unpkg.com/gifenc@1.0.3/dist/gifenc.esm.js"));
 
 const getFrames = makeLazy(() => Promise.all(
     Array.from(
